services/order-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
prometheus-client==0.19.0
pydantic==2.5.2
kafka-python==2.0.2
python-multipart==0.0.6
